refactor(memory,core): move audit content type registration from core to memory plugin
Audit ownership now lives entirely inside the memory plugin:
- removed the core-module buildSearchAPI('_audit') block from server.ts
- added ctx.search.registerContentType with identical schema to memory/index.ts
- table name `bakin_audit` unchanged; memory owns the reindex generator

The memory plugin already exposes /api/plugins/memory/audit and owns
audit reading, so search ownership follows data ownership.
